Refactor logic for returning kids from terminal vs nonterminal fields
