+++ tests/core/structures/acroform/PDFAcroField.spec.ts
@@ -93,29 +93,6 @@ describe('PDFAcroField', () => {
     });
   });
 
-  describe('returns the Kids acrofields', () => {
-    it('when it is defined', () => {
-      const kids = PDFArray.withContext(context);
-      const childDict = PDFDict.fromMapWithContext(
-        new Map([[PDFName.FT, PDFName.Btn]]),
-        context,
-      );
-      const childDictRef = context.register(childDict);
-      const childAcroField = PDFAcroField.fromDict(childDict);
-      kids.push(childDictRef);
-      dict.set(PDFName.Kids, kids);
-      const acroFormFieldDict = PDFDict.fromMapWithContext(dict, context);
-      const field = PDFAcroField.fromDict(acroFormFieldDict);
-      expect(field.getKids()).toEqual([childAcroField]);
-    });
-
-    it('when it is undefined', () => {
-      const acroFormFieldDict = PDFDict.fromMapWithContext(dict, context);
-      const field = PDFAcroField.fromDict(acroFormFieldDict);
-      expect(field.getKids()).toBe(undefined);
-    });
-  });
-
   describe('returns the partial field name', () => {
     it('when it is defined', () => {
       const fieldName = PDFString.of('PersonalData');
